hal: microchip: mec5: BC-Link initialize and start API's
Change the BC-Link transfer start API data parameter from
a pointer to a normal data parameter. The data parameter
is only required when starting a BC-Link write. Data parameter
not requires when starting a BC-Link read operation.
Fix bug that was corrupting the BC-Link control register in the
BC-Link initialization API.

Signed-off-by: Scott Worley <scott.worley@microchip.com>